+++ d00/ex01/utils.h
@@ -6,7 +6,7 @@
 /*   By: ivankozlov <ivankozlov@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/25 15:54:06 by ivankozlov        #+#    #+#             */
-/*   Updated: 2018/06/25 16:11:33 by ivankozlov       ###   ########.fr       */
+/*   Updated: 2019/04/29 18:09:13 by ivankozlov       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,6 @@
 int				getNumber(void);
 std::string		toLower(std::string s);
 std::string		getInfo(std::string field);
-std::string		truncate(std::string s, int len);
+std::string		truncate(std::string s, size_t len);
 
 #endif
